refactor(boylove): separate `Url` into `url.rs`
